Break out AuditInfoFormView into its own file (#2971)
* Break out AuditInfoFormView into its own file.

* Minor corrections for moving AuditInfoFormView into its own file.
